Add FileChangeTracker in core, unit tests now work
