+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.TypeSystem/MonoDevelopMetadataReference.cs
@@ -4,6 +4,7 @@ using System.Diagnostics;
 using System.Diagnostics.Contracts;
 using System.IO;
 using Microsoft.CodeAnalysis;
+using MonoDevelop.FSW;
 
 namespace MonoDevelop.Ide.TypeSystem
 {
@@ -34,13 +35,9 @@ namespace MonoDevelop.Ide.TypeSystem
 			_fileChangeTracker.UpdatedOnDisk += OnUpdatedOnDisk;
 		}
 
-		public string FilePath {
-			get { return _fileChangeTracker.FilePath; }
-		}
+		public string FilePath => _fileChangeTracker.FilePath;
 
-		public MetadataReferenceProperties Properties {
-			get { return _properties; }
-		}
+		public MetadataReferenceProperties Properties => _properties;
 
 		public PortableExecutableReference CurrentSnapshot {
 			get {
@@ -52,10 +49,7 @@ namespace MonoDevelop.Ide.TypeSystem
 			}
 		}
 
-		void OnUpdatedOnDisk(object sender, EventArgs e)
-		{
-			UpdatedOnDisk?.Invoke (this, EventArgs.Empty);
-		}
+		void OnUpdatedOnDisk (object sender, EventArgs e) => UpdatedOnDisk?.Invoke (this, EventArgs.Empty);
 
 		public void Dispose ()
 		{
@@ -68,61 +62,6 @@ namespace MonoDevelop.Ide.TypeSystem
 			_currentSnapshot = new Snapshot (_provider, Properties, FilePath);
 		}
 
-		string GetDebuggerDisplay ()
-		{
-			return Path.GetFileName (FilePath);
-		}
-
-		/// <summary>
-		/// Helper class for working with FileSystemWatcher to observe any change on single file.
-		/// </summary>
-		internal class FileChangeTracker : IDisposable
-		{
-			public event EventHandler UpdatedOnDisk;
-
-			FileSystemWatcher watcher;
-
-			public string FilePath { get; }
-
-			public FileChangeTracker (string filePath)
-			{
-				FilePath = filePath;
-				watcher = new FileSystemWatcher (Path.GetDirectoryName (filePath), Path.GetFileName (filePath)) {
-					NotifyFilter = NotifyFilters.LastWrite | NotifyFilters.Size | NotifyFilters.FileName | NotifyFilters.CreationTime,
-				};
-
-				watcher.Changed += OnChanged;
-				watcher.Created += OnChanged;
-				watcher.Deleted += OnChanged;
-				watcher.Renamed += OnRenamed;
-				watcher.EnableRaisingEvents = true;
-
-				// Currently fails on mono due to https://github.com/mono/mono/issues/8712
-				Debug.Assert (watcher.EnableRaisingEvents);
-			}
-
-
-			void OnRenamed (object sender, RenamedEventArgs e)
-			{
-				UpdatedOnDisk?.Invoke (this, e);
-			}
-
-			void OnChanged (object sender, FileSystemEventArgs e)
-			{
-				UpdatedOnDisk?.Invoke (this, e);
-			}
-
-			public void Dispose ()
-			{
-				if (watcher != null) {
-					watcher.Changed -= OnChanged;
-					watcher.Created -= OnChanged;
-					watcher.Deleted -= OnChanged;
-					watcher.Renamed -= OnRenamed;
-					watcher.Dispose ();
-					watcher = null;
-				}
-			}
-		}
+		string GetDebuggerDisplay () => Path.GetFileName (FilePath);
 	}
 }
