+++ main/tests/Ide.Tests/MonoDevelop.Ide.Tests.csproj
@@ -133,7 +133,6 @@
     <Compile Include="MonoDevelop.Ide.Editor\DocumentReloadTests.cs" />
     <Compile Include="MonoDevelop.Ide.RoslynServices.Options\FullSolutionAnalysisTests.cs" />
     <Compile Include="MonoDevelop.Ide.TypeSystem\MonoDevelopMetadataReferenceManagerTests.cs" />
-    <Compile Include="MonoDevelop.Ide.TypeSystem\FileChangeTrackerTests.cs" />
     <Compile Include="MonoDevelop.Ide.TypeSystem\MonoDevelopMetadataReferenceManager.MetadataReferenceCacheTests.cs" />
   </ItemGroup>
   <ItemGroup>
